+++ example/src/app/pipes/millis-to-iso/millis-to-iso.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { MillisToIsoPipe } from './millis-to-iso.pipe';
+
+describe('MillisToIsoPipe', () => {
+  it('create an instance', () => {
+    const pipe = new MillisToIsoPipe();
+    expect(pipe).toBeTruthy();
+  });
+});