+++ ios/Plugin/CapacitorCalendarPlugin.m
@@ -28,4 +28,8 @@ CAP_PLUGIN(CapacitorCalendarPlugin, "CapacitorCalendar",
            CAP_PLUGIN_METHOD(requestFullCalendarAccess, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(requestFullRemindersAccess, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(requestReadOnlyCalendarAccess, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(modifyEventWithPrompt, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(modifyEvent, CAPPluginReturnNone);
+           CAP_PLUGIN_METHOD(fetchAllCalendarSources, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(fetchAllRemindersSources, CAPPluginReturnPromise);
 )