+++ CHANGELOG.md
@@ -14,6 +14,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Changelogs for the versions supporting Capacitor 6.
 
+## [6.6.0] – 2024-07-26
+
+### Fixed
+
+- Not being able to read event color
+- Not being able to read events timezone
+- Error when creating event
+
+### Added
+
+- Color to calendar object
+- Modify events programmatically
+- Modify events with prompt
+- Method for fetching available calendar sources on iOS
+
+### Removed
+
+- The plugin will not receive updates for capacitor 5
+
 ## [6.5.0] – 2024-05-30
 
 ### Changed